requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Async HTTP
aiohttp>=3.9.1
aioresponses>=0.7.6

# Exchange Integration
ccxt>=4.2.24

# Configuration
pyyaml>=6.0.1

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0

# Database
aiosqlite>=0.19.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
